+++ src/Accounts/Authenticators/ManagedServiceIdentityAuthenticator.cs
@@ -12,6 +12,7 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -30,6 +31,7 @@ namespace Microsoft.Azure.PowerShell.Authenticators
             DefaultMSILoginUri = "http://169.254.169.254/metadata/identity/oauth2/token",
             DefaultBackupMSILoginUri = "http://localhost:50342/oauth2/token";
 
+        private static Regex SystemMsiNameRegex = new Regex(@"MSI@\d+");
 
         public override Task<IAccessToken> Authenticate(AuthenticationParameters parameters, CancellationToken cancellationToken)
         {
@@ -37,9 +39,10 @@ namespace Microsoft.Azure.PowerShell.Authenticators
 
             var scopes = new[] { GetResourceId(msiParameters.ResourceId, msiParameters.Environment) };
             var requestContext = new TokenRequestContext(scopes);
-            ManagedIdentityCredential identityCredential = new ManagedIdentityCredential();
-            var tokenTask = identityCredential.GetTokenAsync(requestContext);
-            return MsalAccessToken.GetAccessTokenAsync(tokenTask, msiParameters.TenantId, msiParameters.Account.Id);
+            var userAccountId = SystemMsiNameRegex.IsMatch(msiParameters.Account.Id) ? null : msiParameters.Account.Id;
+            ManagedIdentityCredential identityCredential = new ManagedIdentityCredential(userAccountId);
+            return MsalAccessToken.GetAccessTokenAsync(identityCredential, requestContext, cancellationToken,
+                msiParameters.TenantId, msiParameters.Account.Id);
         }
 
         public override bool CanAuthenticate(AuthenticationParameters parameters)