+++ src/Resources/ResourceManager/Formatters/WhatIfOperationResultFormatter.cs
@@ -176,7 +176,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Formatters
                 return;
             }
 
-            int scopeCount = resourceChanges.Select(rc => rc.Scope).Distinct().Count();
+            int scopeCount = resourceChanges.Select(rc => rc.Scope.ToUpperInvariant()).Distinct().Count();
 
             this.Builder
                 .AppendLine()
@@ -184,10 +184,10 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Formatters
                 .AppendLine(scopeCount == 1 ? "scope:" : "scopes:");
 
             resourceChanges
-                .OrderBy(rc => rc.Scope)
-                .GroupBy(rc => rc.Scope)
+                .OrderBy(rc => rc.Scope.ToUpperInvariant())
+                .GroupBy(rc => rc.Scope.ToUpperInvariant())
                 .ToDictionary(g => g.Key, g => g.ToList())
-                .ForEach(kvp => FormatResourceChangesInScope(kvp.Key, kvp.Value));
+                .ForEach(kvp => FormatResourceChangesInScope(kvp.Value[0].Scope, kvp.Value));
         }
 
         private void FormatResourceChangesInScope(string scope, IList<PSWhatIfChange> resourceChanges)