+++ src/Synapse/Synapse/Models/SynapseConstants.cs
@@ -32,6 +32,8 @@ namespace Microsoft.Azure.Commands.Synapse.Models
 
         public const string SqlPool = nameof(SqlPool);
 
+        public const string RestorePoint = nameof(RestorePoint);
+
         public const string SqlDatabase = nameof(SqlDatabase);
 
         public const string SparkPoolName = nameof(SparkPoolName);