+++ src/Resources/Resources/ChangeLog.md
@@ -21,6 +21,7 @@
 * Fixed an issue where What-If shows two resource group scopes with different casing
 * Updated `Export-AzResourceGroup` to use the SDK.
 * Added culture info to parse methods
+* Fixed issue where attempts to deploy template specs from a subscription outside of the current subscription context would fail
 
 ## Version 3.0.0
 * Fixed parsing bug